Type SkuRouteObject sites from getSite and rename ssrContext to skuContext.
