+++ fixtures/ssr-sites/src/skuContext.ts
@@ -1,0 +1,12 @@
+import {
+  createSkuContexts,
+  type SiteOf,
+  type SkuRouteObject,
+} from 'sku/runtime';
+
+import type client from './client.js';
+import type server from './server.js';
+
+export const { useSite } = createSkuContexts<typeof server, typeof client>();
+
+export type AppRouteObject = SkuRouteObject<SiteOf<typeof server>>;
