+++ openspec/changes/vite-ssr/specs/ssr/spec.md
@@ -582,28 +582,20 @@ Production remains HTTP.
 
 ### Requirement: Teams can scaffold an SSR app via create
 
-`@sku-lib/create` MUST offer a `ssr` template that MAY omit config `sites` (sku soft-defaults to `'default'`), with `expressTrustProxy: true` in `sku.config`, `routesEntry` configured, a `routes` scaffold with an app-owned pathless root layout at `src/RootLayout.tsx` (optional route-level `sites` only when membership differs), `defineServerEntry` / `defineClientEntry<typeof server>` + `createSkuContexts<typeof server, typeof client>` wiring in `src/ssrContext.ts`, a home page that calls `useSite()`, and realistic default-exported request-entry objects (`middleware`, optional `onListen` / context getters, `onHydrate` — no `routes` re-export, no `Providers`, no `src/App/` shell).
-
-A template with zero or one resolved site MUST omit `getSite` (sku uses the sole resolved site name).
-Multi-site examples MUST declare ≥2 config sites and export `getSite`.
-
-Lazy route page modules in the template MUST use the React Router Data Mode named `Component` export (not `export default`).
-
-The create template id MUST be `ssr` (not `vite-ssr`).
-
-The static `vite` template MUST remain unchanged.
+`@sku-lib/create` MUST offer a template with id `ssr`.
 
 #### Scenario: Create ssr template
 
 - **WHEN** a user runs `@sku-lib/create --template ssr`
 - **THEN** the project is SSR with `expressTrustProxy: true` and `routesEntry` exporting named `routes`
-- **AND** config `sites` may be omitted (soft-default `'default'`) or declare real site names
+- **AND** config `sites` may be omitted or declare real site names
 - **AND** the server entry exports `middleware` (and may export `onListen` / context getters)
 - **AND** the client entry exports `onHydrate` (and may export context getters)
-- **AND** the template wires `createSkuContexts` in `src/ssrContext.ts` and has no `Providers` export
+- **AND** the template wires `createSkuContexts` in `src/skuContext.ts` and has no `Providers` export
 - **AND** the template has `src/RootLayout.tsx` and no `src/App/` directory
 - **AND** the home page calls `useSite()` (and does not use `import.meta.env` for site/environment demo)
 - **AND** a 0–1 site template omits `getSite`
+- **AND** a 0–1 site template keeps unparameterized `SkuRouteObject[]`
 - **AND** request entries do not re-export `routes`
 - **AND** lazy page modules export named `Component`
 - **AND** can `sku start` without further entry setup
@@ -694,11 +686,12 @@ Sku MUST NOT add a runtime experimental gate.
 SSR product docs MUST describe Managed Data Mode vs SSR and the core app contract:
 
 - `routesEntry` + `routes` with optional `sites` membership
+- `SkuRouteObject<SiteOf<typeof server>>` for multi-site `sites` typing (same `Site` as `useSite`)
 - optional `mapRoutePath` for per-site multi-path pages, including index homes via `path: ''` (and correct preload-safe examples)
 - case-sensitive path matching by default (`caseSensitive: true` when omitted) and per-route `caseSensitive: false` opt-out
 - `getSite` tree selection (required when config has >1 site; sole resolved site — soft-default `'default'` when config `sites` is empty — when omitted on 0–1 site)
 - default-exported request entries via `defineServerEntry` / `defineClientEntry<typeof server>` with optional getters and sibling projection
-- always-on `SkuProvider` + `createSkuContexts<typeof server, typeof client>()`
+- always-on `SkuProvider` + `createSkuContexts<typeof server, typeof client>()` in `src/skuContext.ts`
 - optional `middleware` / `onListen` / `onHydrate` / dual-entry `instrumentations` and config `expressTrustProxy`
 - the three value channels vs the app-owned root layout route
 - middleware layers (production: request-context → optional `express.static(publicPath)` when sibling `client/` exists → server-entry `middleware` → HTML, plus the existing `sku start` order)
@@ -717,7 +710,7 @@ Docs MUST NOT tell consumers to install `@vocab/vite` solely so `@vocab/vite/run
 #### Scenario: Primary SSR docs have topic coverage
 
 - **WHEN** a reader opens SSR product docs
-- **THEN** docs cover `routesEntry`, `SkuProvider` / `createSkuContexts`, the three value channels, the app-owned root layout route, named `routes`, optional `sites`, optional `mapRoutePath`, case-sensitive path matching by default with per-route opt-out, `getSite`, `defineServerEntry` / `defineClientEntry<typeof server>`, optional `middleware` / `onListen` / `onHydrate` / `instrumentations`, `expressTrustProxy`, CSP, and response headers
+- **THEN** docs cover `routesEntry`, `SkuProvider` / `createSkuContexts` in `src/skuContext.ts`, the three value channels, the app-owned root layout route, named `routes`, optional `sites`, `SkuRouteObject<SiteOf<typeof server>>`, optional `mapRoutePath`, case-sensitive path matching by default with per-route opt-out, `getSite`, `defineServerEntry` / `defineClientEntry<typeof server>`, optional `middleware` / `onListen` / `onHydrate` / `instrumentations`, `expressTrustProxy`, CSP, and response headers
 - **AND** docs steer page content toward render-time data loading with clients from `useReactContext` / `useClientContext` (not loaders as the default)
 - **AND** docs describe loaders as opt-in for deeply-nested waterfalls, document redirects, response headers, or opt-in `getRouterContext`
 - **AND** docs document optional dual-entry `getRouterContext` and Data Mode vs Framework Mode seeding
@@ -774,6 +767,7 @@ Migrating docs MUST cover:
 
 - named `Component` (not default export) for lazy routes
 - `routesEntry` + `routes` + optional `sites` + `getSite` (required when config has >1 site; fail closed on unknown / non-string site; sole resolved site — soft-default `'default'` when config `sites` is empty — when omitted on 0–1 site)
+- `SkuRouteObject<SiteOf<typeof server>>` for multi-site `sites` typing
 - optional `mapRoutePath` for per-site multi-path pages
 - case-sensitive path matching by default and per-route `caseSensitive: false` opt-out
 - default-exported request-entry objects via `defineServerEntry` / `defineClientEntry<typeof server>` instead of an `onRequest` value return bag; optional `middleware` / `onListen` / `onHydrate` / `instrumentations`
