+++ site/docs/ssr/runtime-api.md
@@ -8,13 +8,14 @@ In the meantime, continue using [Webpack SSR](./webpack-ssr.md).
 Import browser-safe Managed Data Mode helpers from `sku/runtime`.
 Most helpers are documented on the topic pages that use them.
 
-| Export                                                                   | Use for                                                  |
-| ------------------------------------------------------------------------ | -------------------------------------------------------- |
-| [`defineServerEntry` / `defineClientEntry`](./entries.md)                | Typed server and client entry objects                    |
-| [`createSkuContexts`](./providers.md#typed-hooks)                        | Typed `useSite` / `useClientContext` / `useReactContext` |
-| [`usePreloadRoute`](./routing.md#intent-preloading-with-usepreloadroute) | Warm lazy route chunks on intent                         |
-| [`useInsertHtml`](#useinserthtml)                                        | Queue React nodes into the SSR response stream           |
-| [`getCspNonce`](#getcspnonce)                                            | Request a CSP nonce for inline / injected scripts        |
+| Export                                                                            | Use for                                                  |
+| --------------------------------------------------------------------------------- | -------------------------------------------------------- |
+| [`defineServerEntry` / `defineClientEntry`](./entries.md)                         | Typed server and client entry objects                    |
+| [`createSkuContexts`](./providers.md#typed-hooks)                                 | Typed `useSite` / `useClientContext` / `useReactContext` |
+| [`SkuRouteObject` / `SiteOf`](./routing.md#strictly-typed-sites-in-route-objects) | Type `sites` from `getSite`                              |
+| [`usePreloadRoute`](./routing.md#intent-preloading-with-usepreloadroute)          | Warm lazy route chunks on intent                         |
+| [`useInsertHtml`](#useinserthtml)                                                 | Queue React nodes into the SSR response stream           |
+| [`getCspNonce`](#getcspnonce)                                                     | Request a CSP nonce for inline / injected scripts        |
 
 ## `useInsertHtml`
